+++ .eslintignore
@@ -12,9 +12,9 @@ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/bundle.js
 src/cartridges/app_adyen_SFRA/cartridge/client/default/js/checkout/checkoutSFRA5.js
 src/cartridges/app_adyen_SFRA/cartridge/client/default/js/checkout/checkoutSFRA6.js
 src/cartridges/int_adyen_SFRA/cartridge/controllers/middlewares/checkout_services/placeOrder.js
-src/cartridges/int_adyen_SFRA/cartridge/adyen/webhooks
 src/cartridges/int_adyen_SFRA/cartridge/adyen/utils/adyenHelper.js
 src/cartridges/int_adyen_SFRA/cartridge/adyen/scripts/pos/adyenTerminalApi.js
+src/cartridges/int_adyen_webhooks
 
 # Test files
 *.test.js
