+++ src/cartridges/int_adyen_SFRA/cartridge/adyen/scripts/index.js
@@ -14,7 +14,7 @@ const partialPayment = require('*/cartridge/adyen/scripts/partialPayments/partia
 const fetchGiftCards = require('*/cartridge/adyen/scripts/partialPayments/fetchGiftCards');
 const showConfirmationPaymentFromComponent = require('*/cartridge/adyen/scripts/showConfirmation/showConfirmationPaymentFromComponent');
 const showConfirmation = require('*/cartridge/adyen/scripts/showConfirmation/showConfirmation');
-const notify = require('*/cartridge/adyen/webhooks/notify');
+const notify = require('*/cartridge/notify');
 const makeExpressPaymentsCall = require('*/cartridge/adyen/scripts/expressPayments/paypal/makeExpressPaymentsCall');
 const makeExpressPaymentDetailsCall = require('*/cartridge/adyen/scripts/expressPayments/paypal/makeExpressPaymentDetailsCall');
 const saveShopperData = require('*/cartridge/adyen/scripts/expressPayments/paypal/saveShopperData');
