+++ src/cartridges/int_adyen_webhooks/cartridge/checkNotificationAuth.js
@@ -19,7 +19,7 @@
  *  Script used to authentication notification calls from Adyen
  */
 const Site = require('dw/system/Site');
-const AuthenticationUtils = require('*/cartridge/adyen/webhooks/libs/libAuthenticationUtils');
+const AuthenticationUtils = require('*/cartridge/libs/libAuthenticationUtils');
 const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
 
 function check(request) {
