+++ src/cartridges/int_adyen_webhooks/cartridge/int_adyen_webhooks.properties
@@ -1,0 +1,4 @@
+## cartridge.properties for cartridge int_adyen_webhooks
+# Friday, July 11, 2025 15:16:39 GMT+0200 (Central European Summer Time)
+demandware.cartridges.int_adyen_webhooks.multipleLanguageStorefront=true
+demandware.cartridges.int_adyen_webhooks.id=int_adyen_webhooks
