+++ src/cartridges/int_adyen_webhooks/cartridge/libs/libAuthenticationUtils.js
