+++ src/cartridges/int_adyen_webhooks/cartridge/notify.js
@@ -1,7 +1,7 @@
 const Transaction = require('dw/system/Transaction');
 const URLUtils = require('dw/web/URLUtils');
-const checkAuth = require('*/cartridge/adyen/webhooks/checkNotificationAuth');
-const handleNotify = require('*/cartridge/adyen/webhooks/handleNotify');
+const checkAuth = require('*/cartridge/checkNotificationAuth');
+const handleNotify = require('*/cartridge/handleNotify');
 const AdyenConfigs = require('*/cartridge/adyen/utils/adyenConfigs');
 const setErrorType = require('*/cartridge/adyen/logs/setErrorType');
 const AdyenLogs = require('*/cartridge/adyen/logs/adyenCustomLogs');
